+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/combobox/DefaultComboBoxBuilder.java
@@ -175,7 +175,7 @@ public class DefaultComboBoxBuilder<T, C extends JComboBox<T>, B extends ComboBo
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       comboBoxModel.refresh();
     }
   }