+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/ColumnSelectionPanel.java
@@ -214,7 +214,7 @@ final class ColumnSelectionPanel<C> extends JPanel {
     }
 
     @Override
-    public void perform() {
+    public void execute() {
       checkBoxes.stream()
               .filter(Component::isFocusOwner)
               .findAny()