+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDependenciesPanel.java
@@ -161,7 +161,7 @@ public final class EntityDependenciesPanel extends JPanel {
   private final class NavigateRightCommand implements Control.Command {
 
     @Override
-    public void perform() {
+    public void execute() {
       int selectedIndex = tabPane.getSelectedIndex();
       tabPane.setSelectedIndex(selectedIndex == tabPane.getTabCount() - 1 ? 0 : selectedIndex + 1);
       requestSelectedTableFocus();
@@ -171,7 +171,7 @@ public final class EntityDependenciesPanel extends JPanel {
   private final class NavigateLeftCommand implements Control.Command {
 
     @Override
-    public void perform() {
+    public void execute() {
       int selectedIndex = tabPane.getSelectedIndex();
       tabPane.setSelectedIndex(selectedIndex == 0 ? tabPane.getTabCount() - 1 : selectedIndex - 1);
       requestSelectedTableFocus();